fix: guard feedback submit state
